Extract tag helper methods into separate helper
